requirements.txt
slack-sdk>=3.21.0
httpx>=0.27.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
